Refactor | #72 | @YongsHub | OperationDays 공통 dto로 리팩토링
