+++ cakk-api/src/main/java/com/cakk/api/validator/OperationValidator.java
@@ -4,12 +4,12 @@ import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 
 import com.cakk.api.annotation.OperationDay;
-import com.cakk.api.dto.request.shop.OperationDayRequest;
+import com.cakk.common.dto.OperationDays;
 
-public class OperationValidator implements ConstraintValidator<OperationDay, OperationDayRequest> {
+public class OperationValidator implements ConstraintValidator<OperationDay, OperationDays> {
 
 	@Override
-	public boolean isValid(OperationDayRequest value, ConstraintValidatorContext context) {
+	public boolean isValid(OperationDays value, ConstraintValidatorContext context) {
 		if (value.days().size() != value.startTimes().size()) {
 			return false;
 		} else if (value.days().size() != value.endTimes().size()) {
