+++ cakk-common/src/main/java/com/cakk/common/dto/OperationDays.java
@@ -1,19 +1,14 @@
-package com.cakk.api.dto.request.shop;
+package com.cakk.common.dto;
 
 import java.time.LocalTime;
 import java.util.List;
 
-import jakarta.validation.constraints.NotNull;
-
 import com.cakk.common.enums.Days;
 
-public record OperationDayRequest(
+public record OperationDays(
 
-	@NotNull
 	List<Days> days,
-	@NotNull
 	List<LocalTime> startTimes,
-	@NotNull
 	List<LocalTime> endTimes
 ) {
 }
